[FLINK-3489] TableAPI refactoring and cleanup
