+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/api/table/expressions/TreeNode.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.flink.api.table.trees
+package org.apache.flink.api.table.expressions
 
 /**
  * Generic base class for trees that can be transformed and traversed.
