+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/api/table/typeutils/NullAwareComparator.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.flink.api.table.typeinfo
+package org.apache.flink.api.table.typeutils
 
 import org.apache.flink.api.common.typeutils.{CompositeTypeComparator, TypeComparator}
 import org.apache.flink.core.memory.{DataInputView, DataOutputView, MemorySegment}
